Refactor AzureScopeRequestMessage namespace to `Message.Request`
