+++ IRH.Ui/Models/Message/Request/AzureScopeRequestMessage.cs
@@ -2,7 +2,7 @@
 using CommunityToolkit.Mvvm.Messaging.Messages;
 using IRH.Ui.Models.Azure;
 
-namespace IRH.Ui.Models.Message;
+namespace IRH.Ui.Models.Message.Request;
 
 public class AzureScopeRequestMessage : RequestMessage<ObservableCollection<AzureItemControlTemplate>>{
